# docs/apis/account-and-transaction/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Account and Transaction API",
    "description": "Account and Transaction API for Open Banking Connector Web App",
    "version": "web-app/5.1.0"
  },
  "paths": {
    "/aisp/account-access-consent-auth-contexts": {
      "post": {
        "tags": [
          "Account Access Consent Auth Contexts"
        ],
        "summary": "Create account access consent auth context (time-sensitive auth session) and return auth URL",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountAccessConsentAuthContext"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentAuthContextCreateResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentAuthContextCreateResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentAuthContextCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/account-access-consent-auth-contexts/{accountAccessConsentAuthContextId}": {
      "get": {
        "tags": [
          "Account Access Consent Auth Contexts"
        ],
        "summary": "Read account access consent auth context",
        "parameters": [
          {
            "name": "accountAccessConsentAuthContextId",
            "in": "path",
            "description": "ID of AccountAccessConsent AuthContext",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentAuthContextReadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentAuthContextReadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentAuthContextReadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/account-access-consents": {
      "post": {
        "tags": [
          "Account Access Consents"
        ],
        "summary": "Create account access consent",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountAccessConsentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentCreateResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentCreateResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/account-access-consents/{accountAccessConsentId}": {
      "get": {
        "tags": [
          "Account Access Consents"
        ],
        "summary": "Read account access consent",
        "parameters": [
          {
            "name": "accountAccessConsentId",
            "in": "path",
            "description": "ID of AccountAccessConsent",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-include-external-api-operation",
            "in": "header",
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentCreateResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentCreateResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountAccessConsentCreateResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Account Access Consents"
        ],
        "summary": "Delete account access consent",
        "parameters": [
          {
            "name": "accountAccessConsentId",
            "in": "path",
            "description": "ID of AccountAccessConsent",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-include-external-api-operation",
            "in": "header",
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectDeleteResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectDeleteResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Read accounts",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Read accounts",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountAccessConsentAuthContext": {
        "required": [
          "AccountAccessConsentId"
        ],
        "type": "object",
        "properties": {
          "AccountAccessConsentId": {
            "type": "string",
            "format": "uuid"
          },
          "Reference": {
            "type": "string",
            "description": "Optional reference for linking object to something else - e.g. a user ID in the client\r\napplication. This field is not used by Open Banking Connector.",
            "nullable": true
          },
          "CreatedBy": {
            "type": "string",
            "description": "Optional \"created by\" string in local database. Similar to \"modified by\" for mutable fields in local database, this\r\nfield\r\ncan be used to denote authorship.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AccountAccessConsentAuthContextCreateResponse": {
        "type": "object",
        "properties": {
          "AuthUrl": {
            "type": "string",
            "description": "Time-sensitive URL to enable end-user authentication via website or mobile app",
            "nullable": true,
            "readOnly": true
          },
          "AppSessionId": {
            "type": "string",
            "description": "App session ID that can be checked when processing post-auth redirect",
            "nullable": true,
            "readOnly": true
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          },
          "AccountAccessConsentId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "State": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "Id": {
            "type": "string",
            "description": "Unique Open Banking Connector ID (used in local database).",
            "format": "uuid",
            "readOnly": true
          },
          "Created": {
            "type": "string",
            "description": "Created timestamp in local database.",
            "format": "date-time",
            "readOnly": true
          },
          "CreatedBy": {
            "type": "string",
            "description": "Optional \"created by\" string in local database. Similar to \"modified by\" for mutable fields in local database, this\r\nfield\r\ncan be used to denote authorship.",
            "nullable": true,
            "readOnly": true
          },
          "Reference": {
            "type": "string",
            "description": "Optional reference for linking object to something else - e.g. a user ID in the client\r\napplication. This field is not used by Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to create requests."
      },
      "AccountAccessConsentAuthContextReadResponse": {
        "type": "object",
        "properties": {
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          },
          "AccountAccessConsentId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "State": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "Id": {
            "type": "string",
            "description": "Unique Open Banking Connector ID (used in local database).",
            "format": "uuid",
            "readOnly": true
          },
          "Created": {
            "type": "string",
            "description": "Created timestamp in local database.",
            "format": "date-time",
            "readOnly": true
          },
          "CreatedBy": {
            "type": "string",
            "description": "Optional \"created by\" string in local database. Similar to \"modified by\" for mutable fields in local database, this\r\nfield\r\ncan be used to denote authorship.",
            "nullable": true,
            "readOnly": true
          },
          "Reference": {
            "type": "string",
            "description": "Optional reference for linking object to something else - e.g. a user ID in the client\r\napplication. This field is not used by Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Read requests."
      },
      "AccountAccessConsentCreateResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadConsentResponse1"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          },
          "BankRegistrationId": {
            "type": "string",
            "description": "ID of associated BankRegistration object",
            "format": "uuid",
            "readOnly": true
          },
          "ExternalApiId": {
            "type": "string",
            "description": "External (bank) API ID. This should be unique between objects created at the\r\nsame bank but we do not assume global uniqueness between objects created at multiple banks.",
            "nullable": true,
            "readOnly": true
          },
          "ExternalApiUserId": {
            "type": "string",
            "description": "User ID at external API (bank) which may or may not be available via ID token \"sub\" claim. If retrieved from ID\r\ntoken or supplied on object creation, it will be stored here.",
            "nullable": true,
            "readOnly": true
          },
          "AuthContextModified": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "AuthContextModifiedBy": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "Id": {
            "type": "string",
            "description": "Unique Open Banking Connector ID (used in local database).",
            "format": "uuid",
            "readOnly": true
          },
          "Created": {
            "type": "string",
            "description": "Created timestamp in local database.",
            "format": "date-time",
            "readOnly": true
          },
          "CreatedBy": {
            "type": "string",
            "description": "Optional \"created by\" string in local database. Similar to \"modified by\" for mutable fields in local database, this\r\nfield\r\ncan be used to denote authorship.",
            "nullable": true,
            "readOnly": true
          },
          "Reference": {
            "type": "string",
            "description": "Optional reference for linking object to something else - e.g. a user ID in the client\r\napplication. This field is not used by Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to AccountAccessConsent Create requests."
      },
      "AccountAccessConsentRequest": {
        "required": [
          "BankRegistrationId"
        ],
        "type": "object",
        "properties": {
          "TemplateRequest": {
            "$ref": "#/components/schemas/AccountAccessConsentTemplateRequest"
          },
          "ExternalApiRequest": {
            "$ref": "#/components/schemas/OBReadConsent1"
          },
          "BankRegistrationId": {
            "type": "string",
            "description": "Specifies BankRegistration object to use when creating the consent.",
            "format": "uuid"
          },
          "ExternalApiObject": {
            "$ref": "#/components/schemas/ExternalApiConsent"
          },
          "ExternalApiUserId": {
            "type": "string",
            "description": "User ID at external API (bank) which may or may not be available via ID token \"sub\" claim. It can also optionally\r\nbe specified here.",
            "nullable": true
          },
          "Reference": {
            "type": "string",
            "description": "Optional reference for linking object to something else - e.g. a user ID in the client\r\napplication. This field is not used by Open Banking Connector.",
            "nullable": true
          },
          "CreatedBy": {
            "type": "string",
            "description": "Optional \"created by\" string in local database. Similar to \"modified by\" for mutable fields in local database, this\r\nfield\r\ncan be used to denote authorship.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request object used to create an AccountAccessConsent. Includes a UK Open Banking request object\r\nplus information on the bank registration and bank functional API to use for the consent."
      },
...
      "AccountAccessConsentTemplateRequest": {
        "type": "object",
        "properties": {
          "Type": {
            "$ref": "#/components/schemas/AccountAccessConsentTemplateType"
          },
          "Parameters": {
            "$ref": "#/components/schemas/AccountAccessConsentTemplateParameters"
          }
        },
        "additionalProperties": false
      },
...
      "AccountsResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadAccount6"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Account read requests"
      },
...
      "ExternalApiConsent": {
        "type": "object",
        "properties": {
          "ExternalApiId": {
            "type": "string",
            "description": "External (bank) API ID, i.e. ID of object at bank. This should be unique between objects created at the\r\nsame bank but we do not assume global uniqueness between objects created at multiple banks.",
            "nullable": true
          },
          "AuthContext": {
            "$ref": "#/components/schemas/AuthContextRequest"
          }
        },
        "additionalProperties": false
      },
...
      "OBReadConsent1": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadConsent1Data"
          },
          "Risk": {
            "description": "Any object.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request object OBReadConsent1 from UK Open Banking Read-Write Account and Transaction API spec\r\n<a href=\"https://github.com/OpenBankingUK/read-write-api-specs/blob/v3.1.10r5/dist/openapi/account-info-openapi.yaml\" />\r\nv3.1.10r5 <a />. Open Banking Connector will automatically\r\ntranslate <i>from</i> this to an older format for banks supporting an earlier spec version."
      },
...
      "OBReadConsentResponse1": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadConsentResponse1Data"
          },
          "Risk": {
            "description": "Any object.",
            "nullable": true
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "Response object OBReadConsentResponse1 from UK Open Banking Read-Write Account and Transaction API spec\r\n<a href=\"https://github.com/OpenBankingUK/read-write-api-specs/blob/v3.1.10r5/dist/openapi/account-info-openapi.yaml\" />\r\nv3.1.10r5 <a />. Open Banking Connector will automatically\r\ntranslate <i>to</i> this from an older format for banks supporting an earlier spec version."
      },
...
      "ObjectDeleteResponse": {
        "type": "object",
        "properties": {
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}